refactor: #49 Refactor code
어.. 많은 걸 수정했네요,, DB쪽인데 jpa entity column type을 DB랑 맞추었습니다,,,

+++ src/main/java/com/webserver/siatwiki/info/dto/InfoDto.java
@@ -13,16 +13,16 @@ public class InfoDto {
     public static class InfoRequestDto {
         private String type;
         private String content;
-        private int personId;
+        private Long personId;
     }
 
     @Data
     @Builder
     public static class InfoResponseDto {
-        private int id;
+        private Long id;
         private String category;
         private String content;
         private LocalDateTime updateDate;
-        private int personId;
+        private Long personId;
     }
 }
