+++ src/main/java/com/webserver/siatwiki/info/repository/InfoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.webserver.siatwiki.info.entity.Info;
 
 @Repository
-public interface InfoRepository  extends JpaRepository<Info, Integer>{
+public interface InfoRepository  extends JpaRepository<Info, Long>{
 
 }
