+++ src/main/java/com/webserver/siatwiki/person/repository/PersonRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.webserver.siatwiki.person.entity.Person;
 
 @Repository
-public interface PersonRepository extends JpaRepository< Person, Integer> {
+public interface PersonRepository extends JpaRepository< Person, Long> {
 }
